requirements.txt
asyncio==3.4.3
Django==2.1.7
djangorestframework==3.9.1
pkg-resources==0.0.0
pytz==2018.9
